Move subprocessing related utilities to utils.subprocess
